[IMP] product_pricelist_alternative move test setup code to common file
